+++ packages/@molt/command/src/lib/Tex/chain.ts
@@ -1,4 +1,5 @@
-import type { BlockParameters, Node, TableParameters } from './index_.js'
+import type { BlockParameters, ListParameters, Node, TableParameters } from './index_.js'
+import { List } from './index_.js'
 import { Leaf, Table } from './index_.js'
 import { Block } from './index_.js'
 
@@ -11,6 +12,10 @@ export interface BlockBuilder<Chain = null> {
 
   table(rows: Block[][])                                                        : Chain extends null ? BlockBuilder : Chain
   table(builder: ($: TableBuilder) => null|TableBuilder)                        : Chain extends null ? BlockBuilder : Chain
+
+  list(parameters: ListParameters, nodeishes: (string|Block)[])                                                 : Chain extends null ? BlockBuilder : Chain
+  list(nodeishes: (string|Block)[])                                                 : Chain extends null ? BlockBuilder : Chain
+  list(builder: ($: ListBuilder) => null|ListBuilder)                           : Chain extends null ? BlockBuilder : Chain
   
   text(text: string)                                                            : Chain extends null ? BlockBuilder : Chain
 
@@ -23,6 +28,12 @@ export interface TableBuilder {
   rows(...rows: Block[][]): TableBuilder
   headers(headers: string[]): TableBuilder
   header(header: null | string): TableBuilder
+}
+
+export interface ListBuilder {
+  set(parameters: ListParameters): ListBuilder
+  item(node: Block | string): ListBuilder
+  items(...nodes: Block[]): ListBuilder
 }
 
 export interface RootBuilder extends BlockBuilder<RootBuilder> {
@@ -40,7 +51,7 @@ const createBlockBuilder = (params?: { getSuperChain: () => any }): BlockBuilder
   const parentNode = new Block()
   // let parentParameters: = parameters ?? {}
 
-  const builder: BlockBuilder = {
+  const $: BlockBuilder = {
     block: (
       ...args:
         | [BlockParameters, string | null]
@@ -68,11 +79,11 @@ const createBlockBuilder = (params?: { getSuperChain: () => any }): BlockBuilder
         }
       }
 
-      return params?.getSuperChain() ?? builder
+      return params?.getSuperChain() ?? $
     },
     set: (parameters: BlockParameters) => {
       parentNode.setParameters(parameters)
-      return params?.getSuperChain() ?? builder
+      return params?.getSuperChain() ?? $
     },
     table: (rows) => {
       const node =
@@ -80,56 +91,99 @@ const createBlockBuilder = (params?: { getSuperChain: () => any }): BlockBuilder
           ? (rows(createTableBuilder()) as any as BuilderInternal<Table>)._.node
           : new Table(rows)
       parentNode.addChild(node)
-      return params?.getSuperChain() ?? builder
+      return params?.getSuperChain() ?? $
+    },
+    list: (
+      ...args:
+        | [nodeishes: (string | Block)[] | (($: ListBuilder) => null | ListBuilder)]
+        | [parameters: ListParameters, nodeishes: (string | Block)[]]
+    ) => {
+      const parameters = args.length === 1 ? null : args[0]
+      const nodeishes = args.length === 1 ? args[0] : args[1]
+      const node =
+        typeof nodeishes === `function`
+          ? (nodeishes(createListBuilder()) as any as BuilderInternal<List>)._.node
+          : new List(nodeishes.map((_) => (typeof _ === `string` ? new Block(new Leaf(_)) : _)))
+      parentNode.addChild(node)
+      if (parameters) {
+        node.setParameters(parameters)
+      }
+      return params?.getSuperChain() ?? $
     },
     text: (text) => {
       parentNode.addChild(new Leaf(text))
-      return params?.getSuperChain() ?? builder
+      return params?.getSuperChain() ?? $
     },
   }
 
   // Define Internal Methods
-  const builderInternal = builder as any as BuilderInternal<Block>
+  const builderInternal = $ as any as BuilderInternal<Block>
   builderInternal._ = {
     node: parentNode,
   }
 
-  return builder
+  return $
+}
+
+export const createListBuilder = (): ListBuilder => {
+  const parentNode = new List()
+  const $: ListBuilder = {
+    set: (parameters) => {
+      parentNode.setParameters(parameters)
+      return $
+    },
+    item: (nodeish) => {
+      const node = typeof nodeish === `string` ? new Block(new Leaf(nodeish)) : nodeish
+      parentNode.items.push(node)
+      return $
+    },
+    items: (...nodeishes) => {
+      const nodes = nodeishes.map((_) => (typeof _ === `string` ? new Block(new Leaf(_)) : _))
+      parentNode.items.push(...nodes)
+      return $
+    },
+  }
+  // Define Internal Methods
+  const builderInternal = $ as any as BuilderInternal
+  builderInternal._ = {
+    node: parentNode,
+  }
+  return $
 }
 
 const createTableBuilder = (): TableBuilder => {
   const parentNode = new Table()
-  const builder: TableBuilder = {
+  const $: TableBuilder = {
     set: (parameters) => {
       parentNode.setParameters(parameters)
-      return builder
+      return $
     },
     row: (...cells) => {
       const cellsNormalized = cells.map((_) => (typeof _ === `string` ? new Block(new Leaf(_)) : _))
       parentNode.rows.push(cellsNormalized)
-      return builder
+      return $
     },
     rows: (...rows) => {
       parentNode.rows.push(...rows)
-      return builder
+      return $
     },
     headers: (headers) => {
       parentNode.headers = headers
-      return builder
+      return $
     },
     header: (header) => {
       if (header !== null) {
         parentNode.headers.push(header)
       }
-      return builder
+      return $
     },
   }
   // Define Internal Methods
-  const builderInternal = builder as any as BuilderInternal
+  const builderInternal = $ as any as BuilderInternal
   builderInternal._ = {
     node: parentNode,
   }
-  return builder
+  return $
 }
 
 export const createRootBuilder = (parameters?: BlockParameters): RootBuilder => {
@@ -137,9 +191,9 @@ export const createRootBuilder = (parameters?: BlockParameters): RootBuilder => 
 
   builder.render = () => {
     const builderInternal = builder as any as BuilderInternal<Block>
-    return builderInternal._.node.render({
-      maxWidth: parameters?.maxWidth ?? process.stdout.columns ?? 100,
-    }).value
+    const node = builderInternal._.node
+    const block = new Block(parameters ?? {}, node)
+    return block.render({ maxWidth: parameters?.maxWidth ?? process.stdout.columns ?? 100 }).value
   }
 
   return builder
